+++ lib/time.dart
@@ -37,6 +37,8 @@ class Timer {
     return _current >= _limit;
   }
 
+  bool isRunning() => _running;
+
   void start() {
     _current = 0;
     _running = true;
@@ -46,4 +48,7 @@ class Timer {
     _current = 0;
     _running = false;
   }
+
+  /// A value between 0 and 1 indicating the timer progress
+  double get progress => _current / _limit;
 }