+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## next
+- Adding Timer#isRunning method
+- Adding Timer#progress getter
+
 ## 0.13.1
 - Adding Timer utility class
 - Adding `destroyOnFinish` flag for AnimationComponent